added DDqmax, init. Split primites.h in two files. Added tests
