+++ rosdyn_core/test/test.launch
@@ -1,0 +1,110 @@
+
+<launch>
+
+<arg name="valgrind" default="false" />
+<arg name="debug" default="false" />
+
+<group ns="file_and_screen_different_appenders" >
+<rosparam>
+  appenders: [file, screen]
+  levels: [debug, info]
+  file_name: "ciao"
+</rosparam>
+</group>
+
+<group ns="ur10_hw" >
+<rosparam>
+  type: cnr/control/FakeRobotHW
+  appenders: [file, screen]
+  levels: [trace, trace]
+  pattern_layout: "[%5p][%d{HH:mm:ss,SSS}][%50M:%04L][%24c] %m%n"
+  file_name: "ur10_hw"
+  joint_names:
+  - shoulder_pan_joint
+  - shoulder_lift_joint
+  - elbow_joint
+  - wrist_1_joint
+  - wrist_2_joint
+  - wrist_3_joint
+  base_link: base_link
+  tool_link: tool0
+  robot_description_param: /robot_description
+  robot_description_planning_param: /robot_description_planning
+  sampling_period: 0.001
+
+  initial_position: [0, -1.57, 0, 0, 0, 0]
+  fake_controller:
+    type: cnr/control/JointStatePublisher
+    controlled_joint:
+    - shoulder_pan_joint
+    - shoulder_lift_joint
+    - elbow_joint
+    - wrist_1_joint
+    - wrist_2_joint
+    - wrist_3_joint
+</rosparam>
+</group>
+
+<include file="$(find ur_description)/launch/ur10_upload.launch" />
+
+<group ns="robot_description_planning">
+  <rosparam>
+  joint_limits:
+    linear_motor_cursor_joint:
+      has_velocity_limits: true
+      max_velocity: 0.5
+      has_acceleration_limits: true
+      max_acceleration: 14 #1.55
+    elbow_joint:
+      has_velocity_limits: true
+      max_velocity: 1.4
+      has_acceleration_limits: true
+      max_acceleration: 3
+    shoulder_lift_joint:
+      has_velocity_limits: true
+      max_velocity: 1.4
+      has_acceleration_limits: true
+      max_acceleration: 14 #1.55
+    shoulder_pan_joint:
+      has_velocity_limits: true
+      max_velocity: 1.4
+      has_acceleration_limits: true
+      max_acceleration: 14 #1.55
+    wrist_1_joint:
+      has_velocity_limits: true
+      max_velocity: 1.4
+      has_acceleration_limits: true
+      max_acceleration: 14 #1.55
+    wrist_2_joint:
+      has_velocity_limits: true
+      max_velocity: 1.4
+      has_acceleration_limits: true
+      max_acceleration: 14 #1.55
+    wrist_3_joint:
+      has_velocity_limits: true
+      max_velocity: 1.4
+      has_acceleration_limits: true
+      max_acceleration: 3
+  </rosparam>
+</group>
+
+
+<group if="$(arg valgrind)">
+  <arg unless="$(arg debug)" name="launch_prefix" value="" />
+  <arg     if="$(arg debug)" name="launch_prefix"
+      value="valgrind -v --track-origins=yes --tool=memcheck --leak-check=full --show-reachable=yes  --num-callers=20 --track-fds=yes  --log-file=/home/feymann/a.txt" />
+      <test launch-prefix="$(arg launch_prefix)" test-name="rosdyn_core_test" pkg="rosdyn_core" type="rosdyn_core_test">
+      </test>
+</group>
+<group unless="$(arg valgrind)">
+  <arg unless="$(arg debug)" name="launch_prefix" value="" />
+  <arg     if="$(arg debug)" name="launch_prefix" value="gdb --ex run --args" />
+
+<test launch-prefix="$(arg launch_prefix)" test-name="rosdyn_core_test" pkg="rosdyn_core" type="rosdyn_core_test">
+</test>
+
+</group>
+</launch>
+
+
+
